ビルド時のエラーの修正
