+++ app/components/HeaderItinerary.tsx
@@ -1,6 +1,7 @@
 import Link from "next/link";
 import { User } from "@prisma/client";
 import Menu from "./auth/navigation/Menu";
+import Image from "next/image";
 
 type HeaderItineraryProps = {
   currentUser: User | null;
@@ -12,7 +13,7 @@ const HeaderItinerary: React.FC<HeaderItineraryProps> = ({ currentUser }) => {
       <div className="flex justify-between items-center mx-2 h-16">
         <Link href="/memorybook">
           <h1>
-            <img
+            <Image
               src="/logo_itinerary.png"
               alt="国内旅行・海外旅行の旅程表作成しおりアプリ「旅のメモリーブック」"
               width="220"
